+++ Adj/Algebra.hs
@@ -4,3 +4,4 @@ import Adj.Algebra.Semigroupoid as Exports
 import Adj.Algebra.Category as Exports
 import Adj.Algebra.Functor as Exports
 import Adj.Algebra.Morphism as Exports
+import Adj.Algebra.Product as Exports