+++ Adj/Algebra/Product.hs
@@ -1,0 +1,5 @@
+module Adj.Algebra.Product where
+
+infixr 7 :*:
+
+data (:*:) left right = left :*: right